Refactor the `Server` class to prepare the implementation of the Graph Store Protocol (#1743)
- Extract `parseHttpRequest` and `extractAccessToken` into separate Class `SPARQLProtocol`.
- Refactor `processQueryOrUpdate` (and methods below it) to take a `ParsedQuery`. Consolidate query preparation into `parseQuery`, `planQuery` and `setupPlannedQuery`.

Both these changes will make integrating the Graph Store Protocol much easier
