+++ src/util/http/UrlParser.h
@@ -67,6 +67,8 @@ struct Update {
 struct None {
   bool operator==(const None& rhs) const = default;
 };
+
+using Operation = std::variant<Query, Update, None>;
 }  // namespace sparqlOperation
 
 // Representation of parsed HTTP request.
@@ -78,9 +80,7 @@ struct ParsedRequest {
   std::string path_;
   std::optional<std::string> accessToken_;
   ParamValueMap parameters_;
-  std::variant<sparqlOperation::Query, sparqlOperation::Update,
-               sparqlOperation::None>
-      operation_;
+  sparqlOperation::Operation operation_;
 };
 
 // Parse the URL path and the URL query parameters of an HTTP Request target.
